Add next.config.prod.js
